refactor(www): make the modal's footer reusable by moving it to modals folder & not make it specific to add-asset-flow modal
